requirements.txt
aiofiles==0.6.0
aiosqlite==0.16.1
autopep8==1.5.5
cffi==1.14.4
click==7.1.2
databases==0.4.1
Distance==0.1.3
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
imutils==0.5.4
numpy==1.20.0
opencv-python==4.5.1.48
pandas==1.2.1
passlib==1.7.4
Pillow==8.1.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
tqdm==4.56.0
typing-extensions==3.7.4.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
